Добавлена функция подачи заявки на занятие

+++ src/main/java/com/example/ilsan34/clientManager/service/RegistrationForLessonService.java
@@ -2,16 +2,17 @@ package com.example.ilsan34.clientManager.service;
 
 import com.example.ilsan34.clientManager.model.RegistrationForLesson;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface RegistrationForLessonService {
     void addNewRegistrationForLesson(RegistrationForLesson registration);
 
-    RegistrationForLesson getById(Long id);
+    RegistrationForLesson getById(BigInteger id);
 
     void updateRegistrationForLesson(RegistrationForLesson registration);
 
-    void deleteRegistrationForLesson(Long id);
+    void deleteRegistrationForLesson(BigInteger id);
 
     List<RegistrationForLesson> getAll();
 }
